Externalise some methods from the base model into assistant traits
