+++ src/Db/Model/Traits/Static/DirtyAttributesTrait.php
@@ -28,7 +28,7 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace District5\Mondoc\Db\Model\Traits;
+namespace District5\Mondoc\Db\Model\Traits\Static;
 
 use District5\Mondoc\Exception\MondocConfigConfigurationException;
 use District5\Mondoc\Helper\MondocTypes;
@@ -37,7 +37,7 @@ use MongoDB\Model\BSONDocument;
 /**
  * Trait DirtyAttributesTrait.
  *
- * @package District5\Mondoc\Db\Model\Traits
+ * @package District5\Mondoc\Db\Model\Traits\Static
  */
 trait DirtyAttributesTrait
 {
