+++ src/Db/Model/Traits/Static/MondocEncryptedFieldsTrait.php
@@ -28,7 +28,7 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace District5\Mondoc\Db\Model\Traits;
+namespace District5\Mondoc\Db\Model\Traits\Static;
 
 
 use District5\Mondoc\Exception\MondocConfigConfigurationException;
@@ -38,7 +38,7 @@ use District5\Mondoc\MondocConfig;
 /**
  * Trait MondocEncryptedFieldsTrait.
  *
- * @package District5\Mondoc\Db\Model\Traits
+ * @package District5\Mondoc\Db\Model\Traits\Static
  */
 trait MondocEncryptedFieldsTrait
 {
