+++ src/Db/Model/Traits/Static/PresetObjectIdTrait.php
@@ -28,14 +28,14 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace District5\Mondoc\Db\Model\Traits;
+namespace District5\Mondoc\Db\Model\Traits\Static;
 
 use MongoDB\BSON\ObjectId;
 
 /**
  * Trait PresetObjectIdTrait.
  *
- * @package District5\Mondoc\Db\Model\Traits
+ * @package District5\Mondoc\Db\Model\Traits\Static
  */
 trait PresetObjectIdTrait
 {
